backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow
pillow
python-multipart==0.0.6
requests==2.31.0
opencv-python==4.10.0.84
matplotlib==3.8.2
numpy
scipy